backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
solana==0.32.0
solders==0.20.0
python-dotenv==1.0.0
base58==2.1.1
